+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Feature - Add lifecycle hooks for Executor.
+
 0.1.0 (2024-11-16)
 ------------------